timesaver_backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
youtube-transcript-api==0.6.2
pytube==15.0.0
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.31.0